Fix group_by constexpr test on GCC
It doesn't seem to like passing a PMD to invoke() in constexpr, which is weird because I'm sure I've done it a lot in the NanoRange tests.

Anyway, manually writing equivalent lambdas fixes it.

Oh, also add a new test cos I've just realised we can use group_by to easily write unique().