Return the original range in to_range() if possible
For the flow types wrapping an STL range, override the `to_range` functions to return the original range when we are storing it (i.e., when `flow::from` is called with an rvalue).  Also override `to<Container>` to return the stored range when `Container` is the same type as the range.

These changes mean that, for example `flow::from(std::vector{1, 2, 3}).to_vector()` is now just a move or two.