Introduce subflows
Some algorithms require forward ranges. To make it possible to implement these, we'll introduce the notion of "subflows".

A subflow of a flow F is a flow over the same data as F, but with an independent iteration state. It is, in a sense, a "checkpoint" of F at the point the subflow is created.

Subflows are not required to be able to out-live their parent flows, so it's safe for them to refer to the same data as their parent.

Not all flows can create subflows ("pure" input ranges cannot, for example). Those which can are now called *multipass flows*, checkable with the `is_multipass_flow` type trait.