Introduce sized and infinite flows
Some flows, such as those wrapping STL containers, know the number of items they contain in advance. For those, we introduce a new function f.size(), which returns the number of remaining items (in constant time). A flow which has a size() member function is called a *sized flow*.

Conversely, some flows are known to never terminate -- those are called *infinite flows*. To mark a flow as infinite, add a static constexpr bool is_infinite = true static member variable.

Obviously, a flow cannot be both sized and infinite at the same time, but we'll just trust people not to be silly.