Implement size() and is_infinite for zip_with()
A zip_with adaptor is infinite if all of its base flows are infinite.

It is sized if at least one of its bases is sized, and the rest are either sized or infinite. Its size is calculated as the minimum size of all of its bases (taking an infinite flow to have size dist_t::max).