Reimplement comparators to avoid <functional>
Unfortunately it's technically now UB to compare pointers using flow::less, as only std::less is blessed with being allowed to do that, but in practise I don't think that's going to be a big deal.

Also unfortunately, the new bottleneck is #include <memory>, which we need for constexpr std::addressof (because that can't be implemented as constexpr in vanilla C++). But we also need it for any_flow (which uses std::unique_ptr), so I don't think it's the end of the world.

Surprisingly, doing all this makes absolutely no difference to compile times with Clang/libc++, where the before-and-after times are almost identical. But with GCC/libstdc++ it takes the compile time of a simple test program from 0.72s to 0.59s, or around an 18% speedup, so I think that's a win.