Big changes to group_by() and chunk()
Although we had a version group_by() that worked with any flow, which was quite a challenge to write, I wasn't overly happy with it, because of the ease with which you get get dangling references.

This new version of group_by() now requires a multipass flow. It has the advantage that it's "safer" (less chance of dangling references), and inner ranges are themselves now multipass (and often sized).

This also required changes to `chunk()`, which (used to) use `group_by()`. We *could* just use `slide`, which actually works fine, but a specialised chunk adaptor might be slightly more efficient.

Finally, our crappy version of split() now also requires multipass flows. I'm not sure how I feel able this: I think flow::from_istream(is).split('delim') is something that should probably work...