Fix flow::split definition for MSVC
MSVC doesn't like using the decltype of a lambda generic parameter in the definition of another... which I guess is kinda reasonable, although it does work in GCC and Clang.

Anyway, we'll work around it by defining a function object with a real named template parameter type and use that instead.