Remove our maybe<T>, use std::optional instead
Well, okay, we're not removing maybe<T>, we're changing it into either an alias for std::optional or an alias for optional_ref<T>, depending on whether T is an lvalue reference type or not. optional_ref<T> is just a wafer-thin wrapper around a raw pointer that does addressof in its constructor so we don't have to.