Some split improvements
Split is not very good. This makes it slightly less not good, but it's still not very good.

Previously, split assumed that the string started with a non-delimiter, in which case it mostly worked. If the string started with a delimiter, it didn't. Now it does. Mostly.

I still want to replace it with a dedicated split adaptor, but that's... difficult.