Second try at supporting C++20 coroutines
It turns that if you supply the /await switch with MSVC, it gives you
<experimental/coroutine>, but attempting to #include <coroutine> is a
hard error. I don't understand. Anyway, we'll work around this for now by
removing /await from our CMakeLists.txts.

The previous approach was also failing for clang + libstdc++. I don't know
if there's some magic way of making this work (it feels like language+lib
support for coroutines would have to be pretty tied together), but we'll
try and avoid the problem by only checking for <coroutine> if we detect
language support for C++20 coros, and only checking for <exp/coroutine> if
we detect language support for TS coros.

If this *still* doesn't work we'll have to do some compiler-specific things,
or else just give up on TS coros and say they're only supported in C++20 mode.